backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic>=2.11.2,<3.0.0
pydantic-settings>=2.6.1
python-jose[cryptography]==3.3.0
httpx==0.28.1
clerk-backend-api==3.0.3
python-multipart==0.0.20